Move schema_validator.rb file to committee root
