+++ lib/committee/schema_validator.rb
